pith/ppuda-main/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
networkx>=2.4
matplotlib
h5py
einops
tqdm
scikit-learn
scipy
pillow
pycocotools
pydot
